+++ bhp/lims/config.py
@@ -2,4 +2,11 @@
 #
 # Copyright 2018 Botswana Harvard Partnership (BHP)
 
+from Products.Archetypes.utils import DisplayList
+
 PRODUCT_NAME = "bhp.lims"
+
+GENDERS = DisplayList((
+    ('m', ("Male")),
+    ('f', ("Female")),
+))