+++ CHANGES.rst
@@ -4,4 +4,5 @@ Changelog
 1.0.0 (unreleased)
 ------------------
 
-- #1  Hide unused fields from Analysis Request Add form
+- #2 New fields in Analysis Request (and visible in AR Add)
+- #1 Hide unused fields from Analysis Request Add form